switch tun to meh/rust-tun

+++ clash_lib/src/config/internal/config.rs
@@ -309,7 +309,7 @@ pub struct TunConfig {
     pub route_all: bool,
     pub routes: Vec<IpNet>,
     pub gateway: IpNet,
-    pub mtu: Option<i32>,
+    pub mtu: Option<u16>,
     pub so_mark: Option<u32>,
     pub route_table: Option<u32>,
     pub dns_hijack: bool,
